requirements.txt
langchain
PyPDF2
python-dotenv
openai
pdfplumber
typesense
tiktoken
chainlit